Updated shortest path

+++ GraphData.h
@@ -25,7 +25,7 @@ class GraphData {
     GraphData();
     GraphData(std::string data1, std::string data2);
     void insert(Node* newNode);
-    double findDist(int node1, int node2);
+    float findDist(int node1, int node2);
     void BFS(int id);
     cs225::PNG graphVisualizer();
     void drawLines(cs225::PNG * vis);
@@ -41,5 +41,5 @@ class GraphData {
 
     int numNodes_;
     map<int, pair<double, double> > nodes_;
-    map<pair<int, int>, double> edges_;
+    map<pair<int, int>, float> edges_;
 };
